// src/app/transaction-types.ts
export class TransactionTypes {
    transaction_types_id: number;
    transaction_types_name_type_transaction: string; 
    transaction_types_description: string; 
    transaction_types_class_transaction: string; 
    transaction_types_accounting_distribution: string; 
    transaction_types_status: string;
    condominums_id: number;
    created_by: number;
    create_date: Date; 
    last_updated_by: number;
    last_updated_date: Date; 

}